+++ apps/server/src/modules/log-manager/index.ts
@@ -4,9 +4,9 @@ export const LogManager = {
 	async readLogs() {
 		try {
 			await AccessLogService.createIndex();
-			console.log("Successfully initialized log service");
+			console.log("✅  Successfully initialized log service");
 		} catch (error) {
-			console.error("Error initializing log service:", error);
+			console.error("🛑  Error initializing log service:", error);
 		}
 	},
 };