+++ packages/parser/src/types.ts
@@ -38,6 +38,10 @@ export interface CompiledParser {
 	parse: (line: string) => ParsedLogLine;
 }
 
+type PartialRecord<K extends keyof any, T> = {
+	[P in K]?: T;
+};
+
 export type RestFn = (
 	args: Record<SquidFieldType, string>,
-) => Record<SquidFieldType, string>;
+) => PartialRecord<SquidFieldType | string, string>;